attributes: separate the RSA modulus from the public key data
Some PKCS#11 libraries (for instance the PKCS#11 engine in libp11) wants
private key to define a CKA_MODULUS attribute for all RSA private keys -
otherwise thay may crash.

The pkcs11.so engine for openssl, for example, tries to read the
CKA_MODULUS attribute of the private key and its CKA_EXPONENT. If it
does not find the exponent, it tries to find the corresponding public
key to get the exponent from it. But a failure to get the private key
modulus will result in a segfault.

Signed-off-by: Emmanuel Deloget <logout@free.fr>