attributes: add the CKA_LABEL attribute
Some PKCS#11 libraries or engine are making use of the CKA_LABEL
attribute when listing available keys, so let's feed them with this
information. Since the TPM is quite limited w.r.t. metadata that could
be associated to the key, we use the key name to create the label
(essentially giving it the same value as the CKA_ID but shown as an
UTF8 string instead of a binary string).

Signed-off-by: Emmanuel Deloget <logout@free.fr>