attribute: add the CKA_CERTIFICATE_TYPE attribute to certificates
This attribute is required by libp11 to allow certificates to be used.
With this change, it's now possible to take advantage of tpm2-pk11 with
programs that uses certificates on any openssl/libp11 setup.

libp11 expects the following attributes on certificates:

* CKA_CLASS (ok)
* CKA_ID (ok)
* CKA_VALUE (ok)
* CKA_CERTIFICATE_TYPE (missing)

Although it's optionnal, it also try to read CKA_LABEL (this is not
handled in this patch).

This change adds the CKA_CERTIFICATE_TYPE attribute, with value
CKC_X_509.

One can then use cURL with

curl --engine pkcs11 \
	--cert-type ENG -E 'pkcs11:id=$CID' \
	--key-type ENG --key 'pkcs11:id=$KID' \
	--cacert /my/ca.pem \
	https://my.protected.url/

CID is the id of the certificate, KID is the id of the private key (i.e.
it's name).

This has been mainly tested with cURL but should work with some other
softwares.

Signed-off-by: Emmanuel Deloget <logout@free.fr>